+++ CHANGELOG.md
@@ -183,3 +183,4 @@
 * Move `Adj.Program.Maybe` to `Adj.Program.Primitive` module
 * Move `Adj.Program.Progress` to `Adj.Program.Primitive` module
 * Move `Adj.Program.Generation` to `Adj.Program.Primitive` module
+* Define `Adj.Program.Primitive.Boolean` module